requirements.txt
# SentinelAI root dependency manifest (submission artifact)
#
# Purpose:
# - Union of dependencies declared across service-level requirements files.
# - Keeps Docker/Compose behavior unchanged (service Dockerfiles still install their own files).
#
# Safety:
# - This file is intentionally broad and mostly unpinned to reduce cross-service resolver conflicts.
# - For runtime builds, use per-service requirements files. (Docker already does this!)

alembic
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apscheduler>=3.10
authlib==1.6.9
attrs==25.4.0
asyncpg
bcrypt<4.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
cyclopts==4.7.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator
exceptiongroup==1.3.1
fastapi
fastmcp
filelock==3.25.0
fsspec==2026.2.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-genai
grpcio
grpcio-tools
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx
httpx-sse==0.4.3
huggingface-hub
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
jiter==0.13.0
json-repair
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
langchain
langchain-core==1.2.17
langchain-mistralai==1.1.1
langchain-openai==1.1.10
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.14
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib
mcp[cli]
mdurl==0.1.2
more-itertools==10.8.0
neo4j>=5.0.0
nlpaug
numpy
onnx
onnxconverter-common
onnxruntime
openai==2.26.0
openapi-core==0.22.0
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib[bcrypt]>=1.7.4
pathable==0.4.4
peft==0.18.1
pluggy==1.6.0
prometheus-client
protobuf
psycopg
psycopg[binary]
pydantic
pydantic-settings
pydantic_core==2.41.5
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pyperclip==1.11.0
pytest
pytest-asyncio
python-dotenv
python-jose[cryptography]
python-multipart
PyYAML==6.0.3
referencing==0.36.2
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
scikit-learn
seaborn
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy
sqlalchemy[asyncio]
sse-starlette==3.3.2
starlette==0.52.1
stripe==11.4.1
tenacity==9.1.4
tiktoken==0.12.0
tokenizers>=0.20.3
torch
tqdm
transformers==4.46.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn
uvicorn[standard]
wandb
Werkzeug==3.1.6
xxhash==3.6.0
zstandard==0.25.0
